+++ w1_py_parse/schemas/__init__.py
@@ -1,6 +1,7 @@
 from .da_root import DA_ROOT_FIELDS
 from .da_permit import DA_PERMIT_FIELDS
 from .da_field import DA_FIELD_FIELDS
+from .da_field_specific import DA_FIELD_SPECIFIC_FIELDS
 
 SCHEMA_ID_TO_NAME = {
     "01": "DAROOT",
@@ -27,4 +28,5 @@ SCHEMA_FIELDS = {
     "01": DA_ROOT_FIELDS,
     "02": DA_PERMIT_FIELDS,
     "03": DA_FIELD_FIELDS,
+    "04": DA_FIELD_SPECIFIC_FIELDS,
 }